+++ cx-portal/src/features/notification/apiSlice.ts
@@ -31,6 +31,11 @@ export interface NotificationArgsProps {
   notificationTopic: string
 }
 
+export interface NotificationReadType {
+  id: string
+  flag: boolean
+}
+
 export const apiSlice = createApi({
   reducerPath: 'info/notifications',
   baseQuery: fetchBaseQuery(apiBaseQuery()),
@@ -64,9 +69,9 @@ export const apiSlice = createApi({
       // configuration for an individual endpoint, overriding the api setting
       keepUnusedDataFor: 10,
     }),
-    setNotificationRead: builder.mutation<void, string>({
-      query: (id) => ({
-        url: `/api/notification/${id}/read`,
+    setNotificationRead: builder.mutation<void, NotificationReadType>({
+      query: (obj) => ({
+        url: `/api/notification/${obj.id}/read?isRead=${obj.flag}`,
         method: 'PUT',
       }),
     }),